Adds spec.md file and script for generating it
