+++ packages/@tinacms/mdx/src/parse/plate.ts
@@ -15,41 +15,65 @@ See the License for the specific language governing permissions and
 limitations under the License.
 
 */
-import { Position } from './remarkToPlate'
 
+/**
+ * @group _MiscellaneousElement
+ */
 export type RootElement = {
   type: 'root'
   children: BlockElement[]
 }
 
+/**
+ * @group BlockElement
+ */
+export type BlockquoteElement = {
+  type: 'blockquote'
+  children: InlineElement[]
+}
+/**
+ * @group BlockElement
+ */
+export type CodeBlockElement = {
+  type: 'code_block'
+  lang?: string
+  value: string
+  children: [EmptyTextElement]
+}
+/**
+ * @group BlockElement
+ */
 export type HeadingElement = {
   type: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6'
   children: InlineElement[]
 }
-export type ParagraphElement = {
-  type: 'p'
-  children: InlineElement[]
-}
-export type MdxBlockElement = {
-  type: 'mdxJsxFlowElement'
-  name: string | null
-  props: Record<string, unknown>
-  children: [EmptyTextElement]
-}
+
+/**
+ * @group BlockElement
+ */
 export type HrElement = {
   type: 'hr'
   children: [EmptyTextElement]
 }
+/**
+ * @group BlockElement
+ */
 export type HTMLElement = {
   type: 'html'
   value: string
   children: [EmptyTextElement]
 }
+/**
+ * @group InlineElement
+ */
 export type HTMLInlineElement = {
   type: 'html_inline'
   value: string
   children: [EmptyTextElement]
 }
+/**
+ * @group BlockElement
+ */
 export type InvalidMarkdownElement = {
   type: 'invalid_markdown'
   value: string
@@ -57,52 +81,83 @@ export type InvalidMarkdownElement = {
   position?: Position
   children: [EmptyTextElement]
 }
-export type CodeBlockElement = {
-  type: 'code_block'
-  lang?: string
-  value: string
-  children: [EmptyTextElement]
-}
-export type BlockquoteElement = {
-  type: 'blockquote'
-  children: InlineElement[]
-}
+/**
+ * @group ListElements
+ */
+export type List = OrderedListElement | UnorderedListElement
+/**
+ * @group ListElements
+ */
 export type ListItemContentElement = {
   type: 'lic'
   children: LicElement[]
 }
+/**
+ * @group ListElements
+ */
 export type ListItemChildrenElement =
   | ListItemContentElement
   | UnorderedListElement
   | OrderedListElement
 
+/**
+ * @group BlockElement
+ */
 export type ListItemElement = {
   type: 'li'
   children: ListItemChildrenElement[]
 }
+/**
+ * @group BlockElement
+ */
 export type UnorderedListElement = {
   type: 'ul'
   children: ListItemElement[]
 }
+/**
+ * @group BlockElement
+ */
+export type MdxBlockElement = {
+  type: 'mdxJsxFlowElement'
+  name: string | null
+  props: Record<string, unknown>
+  children: [EmptyTextElement]
+}
+/**
+ * @group BlockElement
+ */
 export type OrderedListElement = {
   type: 'ol'
   children: ListItemElement[]
 }
-export type List = OrderedListElement | UnorderedListElement
+/**
+ * @group BlockElement
+ */
+export type ParagraphElement = {
+  type: 'p'
+  children: InlineElement[]
+}
+
+/**
+ * @group BlockElement
+ */
 export type BlockElement =
-  | HeadingElement
-  | ParagraphElement
-  | CodeBlockElement
   | BlockquoteElement
-  | MdxBlockElement
+  | CodeBlockElement
+  | HeadingElement
+  | HrElement
   | HTMLElement
   | ImageElement
-  | UnorderedListElement
-  | OrderedListElement
-  | ListItemElement
-  | HrElement
   | InvalidMarkdownElement
+  | ListItemElement
+  | MdxBlockElement
+  | ParagraphElement
+  | OrderedListElement
+  | UnorderedListElement
 
+/**
+ * @group InlineElement
+ */
 export type MdxInlineElement = {
   type: 'mdxJsxTextElement'
   name: string | null
@@ -110,7 +165,17 @@ export type MdxInlineElement = {
   children: [EmptyTextElement]
 }
 
+/**
+ * @remarks
+ * Used specifically to denote no children, used by
+ * the frontend rich-text editor for void nodes
+ *
+ * @group MiscellaneousElement
+ */
 export type EmptyTextElement = { type: 'text'; text: '' }
+/**
+ * @group InlineElement
+ */
 export type TextElement = {
   type: 'text'
   text: string
@@ -118,6 +183,12 @@ export type TextElement = {
   italic?: boolean
   code?: boolean
 }
+/**
+ * @remarks
+ * It may be beneficial to treat this as a block element
+ *
+ * @group InlineElement
+ */
 export type ImageElement = {
   type: 'img'
   url: string
@@ -125,19 +196,31 @@ export type ImageElement = {
   caption?: string | null
   children: [EmptyTextElement]
 }
+/**
+ * @group InlineElement
+ */
 export type LinkElement = {
   type: 'a'
   url: string
   title?: string | null
   children: InlineElement[]
 }
+/**
+ * @group InlineElement
+ */
 export type BreakElement = {
   type: 'break'
   children: [EmptyTextElement]
 }
 
+/**
+ * @group ListElements
+ */
 export type LicElement = InlineElement
 
+/**
+ * @group InlineElement
+ */
 export type InlineElement =
   | TextElement
   | MdxInlineElement
@@ -145,3 +228,28 @@ export type InlineElement =
   | LinkElement
   | ImageElement
   | HTMLInlineElement
+
+/**
+ * @remarks
+ * Positional information for error reporting
+ *
+ * @group _MiscellaneousElement
+ */
+export type Position = {
+  start: PositionItem
+  end: PositionItem
+}
+
+/**
+ * @remarks
+ * Positional information for error reporting
+ *
+ * @group _MiscellaneousElement
+ */
+export type PositionItem = {
+  line?: number | null
+  column?: number | null
+  offset?: number | null
+  _index?: number | null
+  _bufferIndex?: number | null
+}
